Refactor/ convert form into drawer
