+++ src/apiCalls.js
@@ -1,8 +1,6 @@
 const fetchData = async (type) => {
   try {
-    const response = await fetch(
-      `https://deadstock-api.vercel.app/api/${type}`
-    );
+    const response = await fetch(`http://localhost:3001/api/${type}`);
     const json = await response.json();
     return json;
   } catch (error) {
