+++ src/components/Enlarged/Enlarged.js
@@ -1,16 +1,16 @@
 import React from "react";
 import PropTypes from "prop-types";
 import "./Enlarged.css";
-import { useForm } from "react-hook-form";
+import { TemporaryDrawer } from "../Materials/Drawer"; 
 
 
 
 const Enlarged = ({ pair }) => {
-  const { register, handleSubmit } = useForm();
-  const onSubmit = (formData) => {
- console.log(111111, formData);
-    window.location.href = `mailto:${contact}?subject=${formData.subject}&body=Hey! My name is ${formData.name}. I was interested in buying your ${title} for $${formData.message}. If that works, let's chat. (${formData.email})`;
-  };
+  // const { register, handleSubmit } = useForm();
+  // const onSubmit = (formData) => {
+
+  //   window.location.href = `mailto:${contact}?subject=${formData.subject}&body=Hey! My name is ${formData.name}. I was interested in buying your ${title} for $${formData.message}. If that works, let's chat. (${formData.email})`;
+  // };
 
   const {
     title,
@@ -44,40 +44,7 @@ const Enlarged = ({ pair }) => {
           <p>Price: ${price}.00 USD</p>
         </div>
       </div>
-      <form onSubmit={handleSubmit(onSubmit)} className="form-container">
-        <div className="offer-form">
-          <h4 className="copy-msg">
-            Interested?<br></br> Let's send {user} an offer!
-          </h4>
-          <input
-            {...register("subject")}
-            className="offerInput subject"
-            type="text"
-            placeholder="Subject"
-          />
-          <input
-            {...register("name")}
-            className="offerInput name"
-            type="text"
-            placeholder="Name"
-          />
-          <input
-            {...register("email")}
-            className="offerInput email"
-            type="email"
-            placeholder="Email"
-          />
-          <input
-            {...register("message")}
-            className="offerInput amount"
-            placeholder="Dollar Amount"
-            type="number"
-          />
-        </div>
-        <button className="general-button" type="submit">
-          Submit
-        </button>
-      </form>
+      <TemporaryDrawer contact={contact} title={title}/>
     </div>
   );
 };
